RUB-243: extract file blocks — miner, chainstate, mempool, consensus
- miner_coinbase.go: parseCanonicalTx, canonicalTxWeight, buildCoinbaseTx,
  chooseValidTimestamp, mtpMedian, makeHeaderPrefix, mineHeaderNonce,
  assembleBlockBytes, prevTimestamps, unixNowU64 (miner.go 689→572)
- chainstate_io.go: Save, LoadChainState, ChainStatePath, DevnetGenesis*,
  writeFileAtomic, tempPathFor, allocateAndWriteTemp, writeAndSyncTemp,
  syncDir, decodeHexToBytes*, copyUtxo* (chainstate.go 696→314)
- mempool_stats.go: MempoolStats, MempoolAdmissionCounts, Stats,
  AdmissionCounts, noteAdmissionResult types and methods (mempool.go 610→355)
- block_weight.go: txWeightAndStats, TxWeightAndStats, compactSizeLen,
  addU64, mulU64 (block_basic.go 594→502)
- utxo_u128.go: addU64ToU128WithCode, cmpU128, subU128 (utxo_basic.go 536→498)

No behavior change. All extracted into same-package files.
